hafta_5/requirements.txt
langchain>=0.1.0,<0.3.0
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchainhub>=0.1.14
pydantic>=2.0.0,<3.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0